load/rest/requirements.txt
cassandra-driver
aiohttp
aiohttp_swagger3
pandas
